refactor(actions): board como directorio barril
