+++ src/actions/board/deleteBoard.ts
@@ -1,0 +1,9 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import { requireBoardAccess } from '../shared'
+
+export async function deleteBoard({ boardId }: { boardId: string }): Promise<void> {
+	await requireBoardAccess(boardId)
+	await prisma.board.delete({ where: { id: boardId } })
+}
